+++ messaging-service/internal/application/services/chat/personal_chat_service_test.go
@@ -5,8 +5,8 @@ import (
 	"testing"
 
 	"github.com/chakchat/chakchat-backend/messaging-service/internal/application/publish"
-	"github.com/chakchat/chakchat-backend/messaging-service/internal/application/repository"
-	"github.com/chakchat/chakchat-backend/messaging-service/internal/application/repository/mocks"
+	"github.com/chakchat/chakchat-backend/messaging-service/internal/application/storage/repository"
+	"github.com/chakchat/chakchat-backend/messaging-service/internal/application/storage/repository/mocks"
 	"github.com/chakchat/chakchat-backend/messaging-service/internal/domain/personal"
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/mock"
