+++ messaging-service/internal/application/services/chat/secret_personal_chat_service.go
@@ -7,7 +7,7 @@ import (
 	"github.com/chakchat/chakchat-backend/messaging-service/internal/application/dto"
 	"github.com/chakchat/chakchat-backend/messaging-service/internal/application/publish"
 	"github.com/chakchat/chakchat-backend/messaging-service/internal/application/publish/events"
-	"github.com/chakchat/chakchat-backend/messaging-service/internal/application/repository"
+	"github.com/chakchat/chakchat-backend/messaging-service/internal/application/storage/repository"
 	"github.com/chakchat/chakchat-backend/messaging-service/internal/domain"
 	"github.com/chakchat/chakchat-backend/messaging-service/internal/domain/secpersonal"
 	"github.com/google/uuid"
