+++ messaging-service/internal/application/storage/repository/update_repository.go
@@ -1,0 +1,4 @@
+package repository
+
+type UpdateRepository interface {
+}
