+++ messaging-service/internal/application/storage/transaction_provider.go
@@ -1,0 +1,1 @@
+package storage
